Add applescript
These are some generic AppleScripts. If you use the .applescript file ending, the script will be stored in an uncompiled text-based form, which allows better version control with git. This makes it easier to collaborate on scripts, because everyone can immediately read the code and see which lines have been changed over time